+++ src/Saving.java
@@ -2,7 +2,7 @@ public class Saving extends Account {
     private double interestRate=0.0;
     private Client owner;
     // Constructor with parameters
-    public Saving(int number, Client owner, double interestRate) {
+    public Saving(long number, Client owner, double interestRate) {
         super(number, owner);
         this.owner = owner;
         this.interestRate = interestRate;
